+++ zerver/tests/test_event_system.py
@@ -1004,7 +1004,7 @@ class FetchQueriesTest(ZulipTestCase):
             with mock.patch("zerver.lib.events.always_want") as want_mock:
                 fetch_initial_state_data(user)
 
-        self.assert_length(queries, 35)
+        self.assert_length(queries, 36)
 
         expected_counts = dict(
             alert_words=1,
@@ -1027,7 +1027,7 @@ class FetchQueriesTest(ZulipTestCase):
             realm_linkifiers=1,
             realm_playgrounds=1,
             realm_user=3,
-            realm_user_groups=2,
+            realm_user_groups=3,
             realm_user_settings_defaults=1,
             recent_private_conversations=1,
             starred_messages=1,